requirements.txt
Django==1.9.5
djangorestframework==3.3.3

psycopg2==2.6.1
